feat: refactor note detail page to use separate Content component
